Adding configuration models
Added all models for the configuration and their methods as described in the
documentation. Every model except LevelConfig should already be fully
implemented. LevelConfig still has todos for saving and loading the
configuration as the CryptoForestCryptograph is needed to implement these.
Also added two custom Exceptions not mentioned in the documentation.

Also adds project files, MIT licence and default .gitattributes and .gitignore.