Adding file searching and directory structure
Added the models needed to store a file and directory structure.
The SearchedDirectory contains the properties for storing the directories and files together with SearchedFile.
Added FileSearch using these SearchedDirectories and
SearchedFiles. Implemented the GetStructure in the FileSearch
based on the POC to create the file and directory structure.
How the structure is searched can be defined with the properties.