Improving FileSearch flexibility
Improved the flexibility and understandability of the FileSearch
by adding a new property SpecificFiles. This property allows
to only include the specific file paths. The property IncludePaths
now also serves as a toggle between including and excluding
during exploring the directory structure.

Also removed the unneded FileId prop of SearchedFile and
GetFilePathById method of SearchedDirectory.