Starting the implementation of CryptoForest<T>
Started the implementation of the CryptoForest<T> class and created
the creation methods and the first AddItemAsync method. The other
methods where implemented as stubs.
Also added the EntryExists method to the ICryptoForestStorage and
the file system implementation to check if an entry already exists in
the storage or not.
Added a stub for saving the config in the LevelConfig.