Adding CryptoForest methods for item management
Added methods for managing items with the CryptoForest. These
methods include:
- Adding
- Getting
- Removing
Added the method RemoveEntryAsync to the ICryptoForestStorage
to remove entries in the storage.
Added the method GetLevelOfItem to LevelConfig to get a level
by item GUID as this allows methods in the CryptoForest to only
need the item GUID and not the level GUID.
Also adds two new exceptions that can be thrown if an item or
level with the same key already exists.