Adding README.md description
Added the description of the library to the README.md. As every method
and class is documented in the C# documentation style, which is also
displayed during development in Visual Studio, most of the documentation
is already present and no other README.md is needed as there won't be
any added value.

Also added a method to move an item into another level to the
CryptoForest.