Fixing several issues discovered while testing the GUI
Fixed issue with KeyIV validation where IV length is expected to be the same length as the key but AES256 can only have 128 bits.
Fixed issue with LevelConfig GetLevel and HasLevel methods where the current level was not considered.
Fixed issue with FileSearch not including file paths of the found items making it impossible to encrypt them.
Extended FileSearch with an additional constructor if specific files are searched.
Fixed issue in FileSearch where the recursion would always add the found files and directories to the base directory.