Fixing issues with decrypting files
Fixed the issues during decrypting a file containing multiple files.
The first issue was that the models SearchedDirectory and SearchedFile
didn't contain a constructor without parameters and System.Json couldn't
create an instance of it.
The second issue was that when reading the directory structure data
not everything was read.