Improving RemoveLevelAsync further
Improved the RemoveLevelAsync method from CryptoForest further
to return the removed GUIDs of items and levels instead of
just true or false as this allows to find out which levels
and items were removed.

Also changed the CryptoForestConfig to have public properties
so that System.Text.Json can create a JSON from this model.